# Improve monday.Format() performance by using strings.Builder