The `monday` package is an incredibly useful utility for localizing dates. However, after profiling some code, I've found that `monday.Format` has shown up in some hot loops.  My investigation has found a few areas where it's possible to get a speedup with fairly minimal changes.

For instance, the `Format` function internally relies on string concatenation rather than making use of `strings.Builder` or `bytes.Buffer`; this means that the formatting loops will allocate more often than they would otherwise need to. Additionally, specifying the slice and map capacity has improved performance in a few places, since they similarly require that the slice or map allocate backing arrays, reallocate them, and so forth.

Here are the results of running the new `BenchmarkFormat` method on `master` (labelled old) and my branch (labelled new):

```
name      old time/op    new time/op    delta
Format-8    2.44ms ±11%    2.03ms ±12%  -16.99%  (p=0.000 n=10+9)

name      old alloc/op   new alloc/op   delta
Format-8     665kB ± 0%     593kB ± 0%  -10.82%  (p=0.000 n=10+9)

name      old allocs/op  new allocs/op  delta
Format-8     23.6k ± 0%     20.0k ± 0%  -15.05%  (p=0.000 n=10+8)
```

This is a fairly noticeable decrease in allocations and time taken. Note:  I did not add benchmarks for all of the functions, just `Format`. However, functions like `GetLongDays` will be _much_ faster, improving more than `Format` did.