@ztstewart 2019-07-06
Test results:

```
zstewart@LAPTOP-J9J9VKM0:ztstewart/monday ‹zts/append*›$ go test -race ./...
ok      github.com/ztstewart/monday
```